commit after zadanie3.2 part2
