+++ src/main/java/pl/javastart/task/Test.java
@@ -7,12 +7,11 @@ public class Test {
         Circle circle = new Circle(21);
         Triangle triangle = new Triangle(4, 3, 5);
         Rectangle rectangle = new Rectangle(45, 64.8);
-        ShapeCalculator shapeCalculator = new ShapeCalculator();
 
-        double squareArea = shapeCalculator.calculateSquareArea(square);
-        double circleArea = shapeCalculator.calculateCircleArea(circle);
-        double trianglePerimeter = shapeCalculator.calculateTrianglePerimeter(triangle);
-        double rectPerimeter = shapeCalculator.calculateRectPerimeter(rectangle);
+        double squareArea = square.calculateSquareArea(square);
+        double circleArea = circle.calculateCircleArea(circle);
+        double trianglePerimeter = triangle.calculateTrianglePerimeter(triangle);
+        double rectPerimeter = rectangle.calculateRectPerimeter(rectangle);
         System.out.println("Pole kwadratu wynosi: " + squareArea);
         System.out.println("Pole koła wynosi: " + circleArea);
         System.out.println("Obwód trójkąta wynosi: " + trianglePerimeter);
